requirements.txt
bcrypt==3.2.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
Jinja2==3.0.3
motor==3.4.0
passlib==1.7.4
pydantic==2.7.1
pydantic-settings==2.3.3
pydantic_core==2.18.2
PyJWT==2.3.0
pymongo==4.7.2
python-decouple==3.6
python-dotenv==1.0.1
python-multipart==0.0.9
uvicorn==0.30.0
schedule==1.2.2
python-dateutil==2.8.2
APScheduler==3.10.4